@@ -1,5 +1,3 @@
-import org.gradle.configuration.BuildOperationScriptPlugin
-
 plugins {
     id'java'
     id'application'
@@ -16,10 +14,14 @@ allprojects {
 
     group 'di'
 
-    project.sourceCompatibility = project.targetCompatibility = 14.0
+    project.sourceCompatibility = project.targetCompatibility = 15.0
 
     repositories {
-        jcenter()
+        mavenCentral()
+        maven {
+            url 'https://m2.dv8tion.net/releases'
+            name 'm2-dv8tion'
+        }
     }
 
     jar {
@@ -35,12 +37,14 @@ allprojects {
     }
 
     dependencies {
-        implementation "net.dv8tion:JDA:4.2.0_222"
+        implementation "net.dv8tion:JDA:4.3.0_331"
         implementation "org.fusesource.leveldbjni:leveldbjni-all:1.8"
-        implementation 'com.google.code.gson:gson:2.8.6'
-        implementation 'org.slf4j:slf4j-api:1.7.30'
-        implementation 'org.slf4j:slf4j-simple:1.7.30'
-        implementation 'net.dean.jraw:JRAW:1.1.0'
+        implementation 'com.google.code.gson:gson:2.8.8'
+        implementation 'org.slf4j:slf4j-api:1.7.32'
+        implementation 'org.slf4j:slf4j-simple:1.7.32'
+        implementation files("${project.rootDir}/libs/JRAW-1.1.0.jar")
+        implementation group: 'club.minnced', name: 'discord-webhooks', version: '0.5.8'
+
 
         Project mainBot = project(":modular_bot")
         Project debugBot = project(":module_debugger")
@@ -65,7 +69,8 @@ class MoveFiles extends DefaultTask {
     @TaskAction
     def execute() {
         String path = 'build\\libs\\'
-        project.delete(project.files("Jar files"))
+
+        project.delete(project.files("Jar files\\*"))
 
         for(Project _project : project.subprojects) {
             if(_project == project.project(":modular_bot")) {
