@@ -1,22 +1,35 @@
 package classes.modules;
 
 import classes.Database;
+import classes.Database.InvalidDatabaseException;
 import classes.Tools;
+import core.BotMain;
 import core.Commands;
 import core.Constants;
 import interfaces.IBotModule;
+import net.dv8tion.jda.api.JDA;
+import net.dv8tion.jda.api.entities.TextChannel;
 import net.dv8tion.jda.api.events.GenericEvent;
+import net.dv8tion.jda.api.events.interaction.ButtonClickEvent;
+import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
+import net.dv8tion.jda.api.interactions.commands.build.CommandData;
+import net.dv8tion.jda.api.interactions.commands.build.OptionData;
 
 import java.io.File;
 import java.io.IOException;
+import java.io.PrintWriter;
+import java.io.StringWriter;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.*;
+import java.util.List;
 
 @SuppressWarnings("SameParameterValue")
 public abstract class BotModule implements IBotModule {
+    public static Boolean DEBUG = false;
+
     public String Name;
     public String Description;
     public String JarName;
@@ -24,12 +37,14 @@ public abstract class BotModule implements IBotModule {
     public List<String> Blacklist = new LinkedList<>();
     public Set<Class<? extends GenericEvent>> AllowedEvents = new HashSet<>();
     public Boolean NeedsDB = true;
+    public Map<String, CommandInfo> SlashCommands = new HashMap<>();
+    public Map<String, ThrowingConsumer<ButtonClickEvent>> Buttons = new HashMap<>();
+    public JDA JDA;
 
     private final List<String> DatabaseFiles = new ArrayList<>();
     private final HashMap<String, Database> Databases = new HashMap<>();
 
-
-    public void init(String mainLocation, String moduleName) {
+    public void init(String mainLocation, String moduleName) throws IOException {
         if(NeedsDB) {
             DatabaseFiles.add("default");
 
@@ -50,7 +65,7 @@ public abstract class BotModule implements IBotModule {
                     Databases.put(file, new Database(new File(databasePath.toUri())));
                 } catch (Exception e) {
                     System.out.println("Exception while initializing db");
-                    e.printStackTrace();
+                    throw e;
                 }
             }
         }
@@ -61,10 +76,15 @@ public abstract class BotModule implements IBotModule {
     }
 
     protected Database getDatabase() {
-        return getDatabase("default");
+        try {
+            return getDatabase("default");
+        } catch (InvalidDatabaseException e) {
+            return null;
+        }
     }
 
-    protected Database getDatabase(String file) {
+    protected Database getDatabase(String file) throws InvalidDatabaseException {
+        if(Databases.get(file) == null) throw new InvalidDatabaseException();
         return Databases.get(file);
     }
 
@@ -74,11 +94,18 @@ public abstract class BotModule implements IBotModule {
         }
     }
 
+    public void coreDeInit() throws Exception {
+        deInit();
+        releaseDatabases();
+    }
+
     public void releaseDatabases() throws IOException {
         for (Database db: Databases.values()) {
             db.release();
         }
     }
+
+    protected void deInit(){}
 
     protected void whitelistServer(String serverId) {
         Whitelist.add(serverId);
@@ -92,12 +119,47 @@ public abstract class BotModule implements IBotModule {
         AllowedEvents.add(event);
     }
 
-    protected void SendModuleMessage(GuildMessageReceivedEvent event, String message, String title) {
+    protected void registerCommand(String name, String description, ThrowingConsumer<SlashCommandEvent> process, OptionData... options) {
+        CommandData commandData = new CommandData(name, description);
+        commandData.addOptions(options);
+
+        SlashCommands.put(name, new CommandInfo(commandData, process));
+    }
+
+    protected void registerCommand(CommandData data, ThrowingConsumer<SlashCommandEvent> process) {
+        SlashCommands.put(data.getName(), new CommandInfo(data, process));
+    }
+
+    protected void registerButton(String name, ThrowingConsumer<ButtonClickEvent> process) {
+        Buttons.put(name, process);
+    }
+
+    protected void processSlashCommand(SlashCommandEvent event) {
+        try {
+            String id = event.getName();
+            if (SlashCommands.containsKey(id))
+                SlashCommands.get(id).getSlashCommandProcessor().accept(event);
+        } catch (Exception e) {
+            printExceptionInformation(e);
+        }
+    }
+
+    protected void processButton(ButtonClickEvent event) {
+        try {
+            String id = event.getComponentId();
+            if(Buttons.containsKey(id))
+                Buttons.get(id).accept(event);
+        } catch (Exception e) {
+            printExceptionInformation(e);
+        }
+    }
+
+    protected void sendModuleMessage(GuildMessageReceivedEvent event, String message, String title) {
         Tools.SendMessage(event, message, title);
     }
 
-    protected void SendModuleMessage(GuildMessageReceivedEvent event, String message) {
-        SendModuleMessage(event, message, Name);
+    protected void sendModuleMessage(GuildMessageReceivedEvent event, String message) {
+        sendModuleMessage(event, message, Name);
     }
 
     public void removeGuildData(String guildID) {
@@ -106,20 +168,98 @@ public abstract class BotModule implements IBotModule {
         }
     }
 
-    public void CoreProcess(GenericEvent event) {
+    public void coreProcess(GenericEvent event) {
         String command;
         if(event instanceof GuildMessageReceivedEvent) {
-            command = Commands.ParsePrefix((GuildMessageReceivedEvent) event);
+            command = Commands.parsePrefix((GuildMessageReceivedEvent) event);
         } else {
             command = "";
         }
 
         try {
-            this.Process(event, command);
+            this.process(event, command);
         } catch (Exception e) {
             e.printStackTrace();
+            printExceptionInformation(e);
         }
     }
 
-    protected abstract void Process(GenericEvent event, String command) throws Exception;
+    protected void printExceptionInformation(Exception e) {
+        if(JDA == null) return;
+
+        TextChannel channel = JDA.getTextChannelById(Constants.Module.REPORT_CHANNEL);
+
+        if(channel != null) {
+            StringWriter sw = new StringWriter();
+            PrintWriter pw = new PrintWriter(sw);
+            e.printStackTrace(pw);
+
+            String exceptionStack = sw.toString();
+            if(exceptionStack.length() > 2000)
+                exceptionStack = exceptionStack.substring(0, 1990) + "...";
+
+            String finalExceptionStack = exceptionStack;
+            channel.sendMessage(String.format("Module %s received exception: %s", Name, e.getMessage())).queue(message ->
+                    channel.sendMessage(finalExceptionStack).queue());
+        }
+    }
+
+    public static void staticPrintException(Throwable e) {
+        if(BotMain.JDA == null) return;
+
+        TextChannel channel = BotMain.JDA.getTextChannelById(Constants.Module.REPORT_CHANNEL);
+
+        if(channel != null) {
+            StringWriter sw = new StringWriter();
+            PrintWriter pw = new PrintWriter(sw);
+            e.printStackTrace(pw);
+
+            String exceptionStack = sw.toString();
+            if(exceptionStack.length() > 2000)
+                exceptionStack = exceptionStack.substring(0, 1990) + "...";
+
+            String finalExceptionStack = exceptionStack;
+            channel.sendMessage(String.format("Static Exception:\n%s", e.getMessage())).queue(message ->
+                    channel.sendMessage(finalExceptionStack).queue());
+        }
+    }
+
+    protected abstract void process(GenericEvent event, String command) throws Exception;
+
+    public void coreReadyModule(JDA jda) throws Exception{
+        if(jda != null)
+            JDA = jda;
+        else if(BotMain.JDA != null)
+            JDA = BotMain.JDA;
+
+        readyModule(JDA);
+    }
+
+    public void readyModule(JDA jda) throws Exception {}
+
+    public static class CommandInfo {
+        CommandData data;
+        ThrowingConsumer<SlashCommandEvent> eventConsumer;
+
+        public CommandInfo(CommandData data, ThrowingConsumer<SlashCommandEvent> eventConsumer) {
+            this.data = data;
+            this.eventConsumer = eventConsumer;
+        }
+
+        public CommandData getCommandData() {
+            return data;
+        }
+
+        public ThrowingConsumer<SlashCommandEvent> getSlashCommandProcessor() {
+            return eventConsumer;
+        }
+    }
+
+    @FunctionalInterface
+    public interface ThrowingConsumer<T> {
+        default void accept(final T elem) throws Exception {
+            acceptThrows(elem);
+        }
+        void acceptThrows(T elem) throws Exception;
+    }
 }
