@@ -1,35 +1,45 @@
+import COTDBot.COTDBot;
+import LevelDBReader.LevelDBReader;
 import classes.modules.BotModule;
+import core.BotMain;
 import core.BotTokens;
-import core.Constants;
 import core.Modules;
+import datetime.Datetime;
 import net.dv8tion.jda.api.JDABuilder;
+import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.events.GenericEvent;
+import net.dv8tion.jda.api.events.ReadyEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import net.dv8tion.jda.api.requests.GatewayIntent;
+import net.dv8tion.jda.api.requests.RestAction;
+import net.dv8tion.jda.api.requests.restaction.CommandListUpdateAction;
 import net.dv8tion.jda.api.utils.ChunkingFilter;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
 import java.lang.reflect.Field;
-import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.EnumSet;
-import java.util.LinkedHashMap;
-import java.util.List;
+import java.util.*;
 
 import static core.Constants.Module.DESC_LIMIT;
 import static core.Constants.Module.NAME_LIMIT;
 
 public class Debugger extends ListenerAdapter {
-    static List<BotModule> modules = new ArrayList<>();
     static Path DBPath;
     static LinkedHashMap<String, BotModule> loaded_modules = new LinkedHashMap<>();
 
+    //Add modules here
+    static void InitModules() throws Exception {
+        //addModule(new Datetime());
+        addModule(new COTDBot());
+    }
+
     static {
         try {
+            BotModule.DEBUG = true;
+
             String mainFolder = "ModularBot";
             String path = new File(Debugger.class.getProtectionDomain().getCodeSource().getLocation().toURI()).getPath();
             String mainPath = path.substring(0, path.indexOf(mainFolder) + mainFolder.length());
@@ -39,12 +49,21 @@ public class Debugger extends ListenerAdapter {
             field.setAccessible(true);
 
             if (Files.notExists(DBPath) && !DBPath.toFile().mkdir()) {
-                System.out.println("Couldn't create main folder");
-            } else {
-                //Add modules here
-                addModule(new LevelDBReader());
-                addModule(new Monitoring());
+                try {
+                    Files.createDirectory(DBPath);
 
+                    InitModules();
+
+                    addModule(new LevelDBReader());
+                    field.set(null, loaded_modules);
+                } catch (Exception e) {
+                    e.printStackTrace();
+                    System.out.println("Couldn't create main folder");
+                }
+            } else {
+                InitModules();
+
+                addModule(new LevelDBReader());
                 field.set(null, loaded_modules);
             }
         } catch (Exception e) {
@@ -59,6 +78,7 @@ public class Debugger extends ListenerAdapter {
             builder.setChunkingFilter(ChunkingFilter.ALL);
             builder.addEventListeners(new Debugger());
             builder.build();
+
         } catch (Exception e) {
             System.out.println("Couldn't log in!");
         }
@@ -66,24 +86,68 @@ public class Debugger extends ListenerAdapter {
 
     @Override
     public void onGenericEvent(@NotNull GenericEvent event) {
-        for(BotModule module : modules) {
-            new Thread(() -> module.CoreProcess(event)).start();
+        for(BotModule module : loaded_modules.values()) {
+            new Thread(() -> module.coreProcess(event)).start();
+        }
+    }
+
+    @Override
+    public void onReady(@NotNull ReadyEvent event) {
+        BotMain.JDA = event.getJDA();
+        RestAction.setDefaultFailure(Throwable::printStackTrace);
+
+        for (BotModule module: loaded_modules.values()) {
+            try {
+                module.coreReadyModule(event.getJDA());
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+
+        loadCommands(event);
+        for(int i = 0; i < 5; i++) {
+            try {
+                Datetime module = new Datetime();
+                module.coreReadyModule(event.getJDA());
+                addModule(module);
+            } catch (Exception ignored) {}
+        }
+    }
+
+    private void loadCommands(ReadyEvent event) {
+        Collection<Guild> guilds = event.getJDA().getGuilds();
+        for(Guild guild : guilds) {
+            for(BotModule module : loaded_modules.values()) {
+                if(module.SlashCommands.size() > 0) {
+                    CommandListUpdateAction update = guild.updateCommands();
+                    for(BotModule.CommandInfo command : module.SlashCommands.values()) {
+                        update = update.addCommands(command.getCommandData());
+                    }
+
+                    update.queue();
+                }
+            }
         }
     }
 
     private static void addModule(BotModule module) {
         try {
+            if(loaded_modules.containsKey(module.getClass().getName())) {
+                BotModule oldModule = loaded_modules.get(module.getClass().getName());
+                oldModule.coreDeInit();
+            }
+
             module.init(DBPath.toString(), module.getClass().getName());
 
             String errors = ValidateModule(module);
             if(errors != null) {
                 System.out.printf(errors, module.getClass().getName());
+                return;
             }
 
             loaded_modules.put(module.getClass().getName(), module);
-            modules.add(module);
         } catch (Exception e) {
-            System.out.println(e.toString());
+            e.printStackTrace();
         }
     }
 
